docker_for_regression/regression_model_coordinator/requirements.txt
flask==2.2.3
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
scipy==1.10.1
mlflow==2.2.1
sqlalchemy==2.0.5
pymysql==1.0.3
waitress==2.1.2
requests==2.28.2
python-dotenv==1.0.0
werkzeug==2.2.3
azure-storage-blob==12.14.1 